+++ misc/multistream-select/src/negotiated.rs
@@ -61,12 +61,12 @@ where
         match Negotiated::poll(Pin::new(&mut io), cx) {
             Poll::Pending => {
                 self.inner = Some(io);
-                return Poll::Pending
+                Poll::Pending
             },
             Poll::Ready(Ok(())) => Poll::Ready(Ok(io)),
             Poll::Ready(Err(err)) => {
                 self.inner = Some(io);
-                return Poll::Ready(Err(err));
+                Poll::Ready(Err(err))
             }
         }
     }
@@ -104,9 +104,8 @@ impl<TInner> Negotiated<TInner> {
 
         let mut this = self.project();
 
-        match this.state.as_mut().project() {
-            StateProj::Completed { .. } => return Poll::Ready(Ok(())),
-            _ => {}
+        if let StateProj::Completed { .. } = this.state.as_mut().project() {
+             return Poll::Ready(Ok(()));
         }
 
         // Read outstanding protocol negotiation messages.
@@ -189,12 +188,9 @@ where
         -> Poll<Result<usize, io::Error>>
     {
         loop {
-            match self.as_mut().project().state.project() {
-                StateProj::Completed { io } => {
-                    // If protocol negotiation is complete, commence with reading.
-                    return io.poll_read(cx, buf)
-                },
-                _ => {}
+            if let StateProj::Completed { io } = self.as_mut().project().state.project() {
+                // If protocol negotiation is complete, commence with reading.
+                return io.poll_read(cx, buf);
             }
 
             // Poll the `Negotiated`, driving protocol negotiation to completion,
@@ -220,12 +216,9 @@ where
         -> Poll<Result<usize, io::Error>>
     {
         loop {
-            match self.as_mut().project().state.project() {
-                StateProj::Completed { io } => {
-                    // If protocol negotiation is complete, commence with reading.
-                    return io.poll_read_vectored(cx, bufs)
-                },
-                _ => {}
+            if let StateProj::Completed { io } = self.as_mut().project().state.project() {
+                // If protocol negotiation is complete, commence with reading.
+                return io.poll_read_vectored(cx, bufs)
             }
 
             // Poll the `Negotiated`, driving protocol negotiation to completion,