+++ core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Update `multihash`.
 
+- Update `multistream-select`.
+
 # 0.24.0 [2020-11-09]
 
 - Remove `ConnectionInfo` trait and replace it with `PeerId`