+++ core/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.25.0 [unreleased]
 
 - The `NetworkConfig` API is now a builder that moves `self`.
+  [PR 1848](https://github.com/libp2p/rust-libp2p/pull/1848/).
 
 - New configurable connection limits for established connections and
   dedicated connection counters. Removed the connection limit dedicated